+++ packages/expo-camera/android/src/main/java/expo/modules/camera/utils/ExifTags.kt
@@ -1,4 +1,4 @@
-package expo.modules.camera.legacy
+package expo.modules.camera.utils
 
 import androidx.exifinterface.media.ExifInterface